pyproject.toml
[tool.poetry]
name = "test-client-api"
version = "0.1.0"
description = ""
authors = ["Oleksandr Kondratiuk <alex_kondr@outlook.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "test_client_api"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
requests = "^2.28.2"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
localstoragepy = "^0.2.3"
selenium = "^4.8.3"
brython = "3.10"
naked = "^0.1.32"
pydantic = {extras = ["email"], version = "^1.10.7"}
requests-toolbelt = "^0.10.1"
cloudinary = "^1.32.0"
pydentic = "^0.0.1.dev3"
psycopg2-binary = "^2.9.6"
sqlalchemy = "^2.0.16"
alembic = "^1.11.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
redis = "^4.5.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
